Version 0.7.14 * separated UI and backend logic by adding ExecutionManager TerminalService and TerminalSession * removed redundant SequenceExecutionTask.java * added restart button for SingularExecutions and Sequences * added rename field for execution tabs * organized exceptions

+++ src/main/java/org/ashot/shellflow/utils/Utils.java
@@ -7,7 +7,7 @@ public class Utils {
     private Utils() {
     }
 
-    public static int calculateDelay(int multiplier, int delayPerCmd) {
+    public static long calculateDelay(int multiplier, long delayPerCmd) {
         if (delayPerCmd == 0) {
             delayPerCmd = 1;
         }
